refactor: address reviewer feedback on PR #181
- Replace `db.repos.notification_bots._db` double-private chain with
  `db._db` in `_check_notification_bot`
- Translate Russian inline comments in test_cli.py to English
- Move `cli_db` and `cli_env` fixtures from test_cli.py /
  test_cli_analytics.py into tests/conftest.py to eliminate duplication
- Remove unused `pytestmark`, `patch`, `pytest` imports from
  test_cli_analytics.py now that fixtures come from conftest
- Add sync comment to `_TG_CHECKS_AFTER_POOL` constant
