+++ tests/conftest.py
@@ -1,9 +1,11 @@
 from __future__ import annotations
 
+import asyncio
 import os
 from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Callable, Collection, Mapping
+from unittest.mock import patch
 
 import pytest
 from telethon import TelegramClient
@@ -42,6 +44,28 @@ async def db():
     await database.initialize()
     yield database
     await database.close()
+
+
+@pytest.fixture
+def cli_db(tmp_path):
+    """Sync fixture: real SQLite for CLI tests."""
+    db_path = str(tmp_path / "cli_test.db")
+    database = Database(db_path)
+    asyncio.run(database.initialize())
+    yield database
+    asyncio.run(database.close())
+
+
+@pytest.fixture
+def cli_env(cli_db):
+    """Patch runtime.init_db to return real db without loading config.yaml."""
+    config = AppConfig()
+
+    async def fake_init_db(config_path: str):
+        return config, cli_db
+
+    with patch("src.cli.runtime.init_db", side_effect=fake_init_db):
+        yield cli_db
 
 
 @pytest.fixture
